backend-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML Libraries
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
python-docx==1.1.0
mammoth==1.6.0
reportlab==4.0.7

# Database
psycopg2-binary==2.9.9
pgvector==0.2.3
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
requests==2.31.0
aiofiles==23.2.1
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0
